+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## UNRELEASE
 
+## 2.10.0
+
+- **Fix** : fix order not updated due to conflict with WPML plugin
+- **Add** : Added Apple Pay (Classic/Blocks Checkout)
+
 ## 2.9.2
 
 - **Fix** : Alma issues on Woocommerce